+++ samples/subsys/console/echo/prj.conf
@@ -1,4 +1,4 @@
-CONFIG_CONSOLE_PULL=y
+CONFIG_CONSOLE_SUBSYS=y
 CONFIG_CONSOLE_GETCHAR=y
 CONFIG_CONSOLE_GETCHAR_BUFSIZE=64
 CONFIG_CONSOLE_PUTCHAR_BUFSIZE=512
