+++ samples/subsys/console/getline/prj.conf
@@ -1,2 +1,2 @@
-CONFIG_CONSOLE_PULL=y
+CONFIG_CONSOLE_SUBSYS=y
 CONFIG_CONSOLE_GETLINE=y
