+++ subsys/shell/shell.c
@@ -14,23 +14,11 @@
 #include <string.h>
 #include <version.h>
 
-#include <console/console.h>
+#include <console.h>
+#include <drivers/console/console.h>
 #include <misc/printk.h>
 #include <misc/util.h>
 #include "mgmt/serial.h"
-
-#ifdef CONFIG_UART_CONSOLE
-#include <console/uart_console.h>
-#endif
-#ifdef CONFIG_TELNET_CONSOLE
-#include <console/telnet_console.h>
-#endif
-#ifdef CONFIG_NATIVE_POSIX_CONSOLE
-#include <console/native_posix_console.h>
-#endif
-#ifdef CONFIG_WEBSOCKET_CONSOLE
-#include <console/websocket_console.h>
-#endif
 
 #include <shell/shell.h>
 
@@ -631,19 +619,8 @@ void shell_init(const char *str)
 	k_thread_create(&shell_thread, stack, STACKSIZE, shell, NULL, NULL,
 			NULL, K_PRIO_COOP(7), 0, K_NO_WAIT);
 
-	/* Register serial console handler */
-#ifdef CONFIG_UART_CONSOLE
-	uart_register_input(&avail_queue, &cmds_queue, completion);
-#endif
-#ifdef CONFIG_TELNET_CONSOLE
-	telnet_register_input(&avail_queue, &cmds_queue, completion);
-#endif
-#ifdef CONFIG_NATIVE_POSIX_STDIN_CONSOLE
-	native_stdin_register_input(&avail_queue, &cmds_queue, completion);
-#endif
-#ifdef CONFIG_WEBSOCKET_CONSOLE
-	ws_register_input(&avail_queue, &cmds_queue, completion);
-#endif
+	/* Register console handler */
+	console_register_line_input(&avail_queue, &cmds_queue, completion);
 }
 
 /** @brief Optionally register an app default cmd handler.
